requirements.txt
Flask==2.3.2
openai==0.27.0
langchain==0.0.148
Werkzeug==2.3.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==20.1.0

unstructured==0.6.7
tiktoken
sentence-transformers==2.2.2
faiss-cpu==1.7.3

cachetools
